// constants/controller.messages.d.ts
declare class ControllerMessages {
    private name;
    CREATE: {
        req: string;
        res: string;
    };
    FETCH: {
        req: string;
        res: string;
    };
    FETCHALL: {
        req: string;
        res: string;
    };
    UPDATE: {
        req: string;
        res: string;
    };
    DELETE: {
        req: string;
        res: string;
    };
    constructor(name: string);
}
export { ControllerMessages };
//# sourceMappingURL=controller.messages.d.ts.map